🐛 fix:  未参加のときにメッセージが表示されない問題を修正
#74
